SCVB0の実装 (#1)
* impl scvb0

* optimization alpha

* fix optfunc

* add minibatch

+++ src/sampler.h
@@ -6,32 +6,37 @@
 
 using namespace std;
 
-namespace Sampler {
+namespace Sampler
+{
 
-inline void init() {
+inline void init()
+{
 	srand(1);
 } //time(NULL));}
 
 // uniform sample Mult(1/K), result [0, K-1]
-inline int uni_sample(int K) {
+inline int uni_sample(int K)
+{
 	int k = rand() % K;
 	return k;
 }
 
 // sample from [0,1)
-inline double uni_sample() {
-	double t = rand() % 10000 / 10000;
+inline double uni_sample()
+{
+	double t = (double)(rand() % 10000) / 10000;
 	return t;
 }
 
 // sample from Mult(p)
-inline int mult_sample(vector<double> p) {
+inline int mult_sample(vector<double> p)
+{
 	int K = p.size();
 	for (int i = 1; i < K; i++)
 		p[i] += p[i - 1];
 
 	double u = double(rand()) / RAND_MAX;
-	int k;  						// record sampled index
+	int k; // record sampled index
 	for (k = 0; k < K; k++)
 		if (p[k] >= u * p[K - 1])
 			break;
@@ -43,15 +48,17 @@ inline int mult_sample(vector<double> p) {
 }
 
 // p is p(1)
-inline bool Bern_sample(float p) {
+inline bool Bern_sample(float p)
+{
 	double u = double(rand()) / RAND_MAX;
 	return (u < p);
 }
 
 // return counts sampling from multinormial distribution p
 // the counts sum to N
-inline void systematic_sample(const vector<double>& p, int N,
-		vector<int>& counts) {
+inline void systematic_sample(const vector<double> &p, int N,
+							  vector<int> &counts)
+{
 	counts.resize(p.size(), 0);
 
 	vector<double> u(N);
@@ -62,8 +69,10 @@ inline void systematic_sample(const vector<double>& p, int N,
 	int i = 0;
 	double s1 = 0;
 	double s2 = p[0];
-	for (int n = 0; n < N - 1; ++n) {
-		while (i < N && u[i] < s2) {
+	for (int n = 0; n < N - 1; ++n)
+	{
+		while (i < N && u[i] < s2)
+		{
 			++i;
 			++counts[n];
 		}
@@ -74,5 +83,5 @@ inline void systematic_sample(const vector<double>& p, int N,
 
 	counts[N - 1] = N - i;
 }
-}
+} // namespace Sampler
 #endif
